+++ ghg_lulc/components/raster_artifacts.py
@@ -3,13 +3,18 @@ from typing import Dict, Tuple
 
 import numpy as np
 import pandas as pd
-from climatoology.base.artifact import Chart2dData, RasterInfo, Artifact, ContinuousLegendData, ArtifactMetadata, Legend
-from climatoology.base.artifact_creators import create_raster_artifact, create_table_artifact, create_chart_artifact
+from climatoology.base.artifact import RasterInfo, Artifact, ContinuousLegendData, ArtifactMetadata, Legend
+from climatoology.base.artifact_creators import create_raster_artifact
 from climatoology.base.computation import ComputationResources
 from climatoology.utility.lulc import LabelDescriptor
 from pydantic_extra_types.color import Color
 
-from ghg_lulc.utils import PROJECT_DIR, GhgStockSource, RASTER_NO_DATA_VALUE, EMISSION_PER_PIXEL_FACTOR, Topics
+from ghg_lulc.components.utils import (
+    PROJECT_DIR,
+    RASTER_NO_DATA_VALUE,
+    EMISSION_PER_PIXEL_FACTOR,
+    Topics,
+)
 
 
 def create_classification_artifacts(
@@ -20,8 +25,8 @@ def create_classification_artifacts(
 ) -> Tuple[Artifact, Artifact]:
     # Hack due to https://gitlab.gistools.geog.uni-heidelberg.de/climate-action/web-app/-/issues/114
     unknown_color = Color('gray')
-    colormap = lulc_before.colormap
-    colormap[0] = unknown_color.as_rgb_tuple()
+    lulc_before.colormap[0] = unknown_color.as_rgb_tuple()
+    lulc_after.colormap[0] = unknown_color.as_rgb_tuple()
     legend = {v.name: Color(v.color) for _, v in labels.items()}
     legend['unknown'] = unknown_color
 
@@ -35,15 +40,10 @@ def create_classification_artifacts(
         tags={Topics.MAPS},
     )
     lulc_before_artifact = create_raster_artifact(
-        raster_info=RasterInfo(
-            data=lulc_before.data.filled(fill_value=RASTER_NO_DATA_VALUE),
-            crs=lulc_before.crs,
-            transformation=lulc_before.transformation,
-            colormap=colormap,
-        ),
+        raster_info=lulc_before,
         legend=Legend(legend_data=legend),
-        resources=resources,
         metadata=lulc_before_metadata,
+        resources=resources,
     )
 
     lulc_after_metadata = ArtifactMetadata(
@@ -55,18 +55,13 @@ def create_classification_artifacts(
         ),
         tags={Topics.MAPS},
     )
-    lulc_after_artifact = create_raster_artifact(
-        raster_info=RasterInfo(
-            data=lulc_after.data.filled(fill_value=RASTER_NO_DATA_VALUE),
-            crs=lulc_after.crs,
-            transformation=lulc_after.transformation,
-            colormap=colormap,
-        ),
-        legend=Legend(legend_data=legend),
-        resources=resources,
-        metadata=lulc_after_metadata,
-    )
 
+    lulc_after_artifact = create_raster_artifact(
+        raster_info=lulc_after,
+        legend=Legend(legend_data=legend),
+        metadata=lulc_after_metadata,
+        resources=resources,
+    )
     return lulc_before_artifact, lulc_after_artifact
 
 
@@ -172,105 +167,3 @@ def patch_change_data(change_data: np.ndarray, orig_colormap: Dict[Number, Tuple
 
     patched_data = patched_data.astype(np.uint8, copy=False)
     return patched_data, patched_colormap
-
-
-def create_summary_artifact(summary_df: pd.DataFrame, resources: ComputationResources) -> Artifact:
-    summary_metadata = ArtifactMetadata(
-        name='Summary of results',
-        filename='summary',
-        summary='Gross emissions, gross sinks, and net emissions/sinks in the analysis period.',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/10a_summary.md').read_text(encoding='utf-8'),
-        tags={Topics.TABLES},
-    )
-    summary_artifact = create_table_artifact(
-        data=summary_df,
-        resources=resources,
-        metadata=summary_metadata,
-    )
-    return summary_artifact
-
-
-def create_area_info_artifact(area_info_df: pd.DataFrame, resources: ComputationResources) -> Artifact:
-    area_info_metadata = ArtifactMetadata(
-        name='Summary of detected changes',
-        filename='area_info',
-        summary='Size of the area classified as a carbon source, carbon sink, and total LULC change during the period '
-        'of analysis.',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/10b_area_info.md').read_text(encoding='utf-8'),
-        tags={Topics.TABLES},
-    )
-    summary_artifact = create_table_artifact(
-        data=area_info_df,
-        resources=resources,
-        metadata=area_info_metadata,
-    )
-    return summary_artifact
-
-
-def create_stock_artifact(
-    stock_df: pd.DataFrame, stock_source: GhgStockSource, resources: ComputationResources
-) -> Artifact:
-    stock_metadata = ArtifactMetadata(
-        name='Carbon stock values per class',
-        filename='stock',
-        summary=f'Carbon stock values for each class according to: {stock_source.value}',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/08_ghg_stocks.md').read_text(encoding='utf-8'),
-        tags={Topics.TABLES},
-    )
-    stock_artifact = create_table_artifact(
-        data=stock_df,
-        resources=resources,
-        metadata=stock_metadata,
-    )
-    return stock_artifact
-
-
-def create_change_type_artifact(change_type_table: pd.DataFrame, resources: ComputationResources) -> Artifact:
-    change_type_metadata = ArtifactMetadata(
-        name='Change areas and carbon flows by LULC change type',
-        filename='stats_change_type',
-        summary='Total change area by LULC change type (ha) and total carbon flows by '
-        'LULC change type (tonnes) in the analysis period.',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/09_stats_change_type.md').read_text(
-            encoding='utf-8'
-        ),
-        tags={Topics.TABLES},
-    )
-    change_type_table_artifact = create_table_artifact(
-        data=change_type_table,
-        resources=resources,
-        metadata=change_type_metadata,
-    )
-    return change_type_table_artifact
-
-
-def create_area_plot_artifact(area_data: Chart2dData, resources: ComputationResources) -> Artifact:
-    area_data_metadata = ArtifactMetadata(
-        name='Change areas by LULC change type (ha)',
-        filename='area_plot',
-        summary='Change areas by LULC change type (ha) in the analysis period.',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/07_area_plot.md').read_text(encoding='utf-8'),
-        tags={Topics.CHARTS},
-    )
-    area_data_artifact = create_chart_artifact(
-        data=area_data,
-        resources=resources,
-        metadata=area_data_metadata,
-    )
-    return area_data_artifact
-
-
-def create_emission_plot_artifact(emission_data: Chart2dData, resources: ComputationResources) -> Artifact:
-    emission_data_metadata = ArtifactMetadata(
-        name='Carbon flows by LULC change type (tonnes)',
-        filename='emission_plot',
-        summary='Carbon flows by LULC change type (tonnes) in the analysis period.',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/06_emission_plot.md').read_text(encoding='utf-8'),
-        tags={Topics.CHARTS},
-    )
-    emission_data_artifact = create_chart_artifact(
-        data=emission_data,
-        resources=resources,
-        metadata=emission_data_metadata,
-    )
-    return emission_data_artifact