+++ ghg_lulc/core/operator_worker.py
@@ -4,7 +4,6 @@ from pathlib import Path
 from typing import List, Tuple
 
 import geopandas as gpd
-import pandas as pd
 import shapely
 from climatoology.base.artifact import RasterInfo
 from climatoology.base.baseoperator import BaseOperator, Artifact, AoiProperties
@@ -21,19 +20,14 @@ from climatoology.utility.lulc import LulcUtility, LulcWorkUnit, FusionMode
 from climatoology.base.exception import ClimatoologyUserError
 from pydantic import HttpUrl
 
-from ghg_lulc.artifact import (
-    create_area_plot_artifact,
+from ghg_lulc.components.raster_artifacts import (
     create_change_artifacts,
-    create_change_type_artifact,
     create_classification_artifacts,
-    create_emission_plot_artifact,
-    create_stock_artifact,
-    create_summary_artifact,
-    create_area_info_artifact,
 )
-from ghg_lulc.emissions import EmissionCalculator
-from ghg_lulc.input import ComputeInput
-from ghg_lulc.utils import (
+from ghg_lulc.components.emissions import EmissionCalculator
+from ghg_lulc.components.statistic_artifacts import create_table_artifacts, create_chart_artifacts
+from ghg_lulc.core.input import ComputeInput
+from ghg_lulc.components.utils import (
     PROJECT_DIR,
     calc_emission_factors,
     fetch_lulc,
@@ -221,56 +215,3 @@ class GHGEmissionFromLULC(BaseOperator[ComputeInput]):
         lulc_after = fetch_lulc(self.lulc_utility, area_after, aoi)
 
         return lulc_before, lulc_after
-
-
-def create_table_artifacts(
-    emission_calculator: EmissionCalculator,
-    emissions_df: gpd.GeoDataFrame,
-    ghg_stock: pd.DataFrame,
-    aoi: shapely.MultiPolygon,
-    params: ComputeInput,
-    resources: ComputationResources,
-) -> List[Artifact]:
-    """
-    Contains the methods to create the table artifacts.
-
-    :param emission_calculator: Class containing the emission estimation methods
-    :param emissions_df: Geodataframe with LULC change polygons and emissions [t] for each change type
-    :param ghg_stock: Dataframe with LULC classes, their GHG stocks and additional info
-    :param params: Operator input
-    :param resources: Ephemeral computation resources
-    :return: List of the table artifacts
-    """
-    ghg_stock_df = EmissionCalculator.filter_ghg_stock(ghg_stock)
-    stock_artifact = create_stock_artifact(ghg_stock_df, params.carbon_stock_source, resources)
-
-    emission_info_df, area_info_df = emission_calculator.summary_stats(emissions_df, aoi)
-    summary_artifact = create_summary_artifact(emission_info_df, resources)
-    area_info_artifact = create_area_info_artifact(area_info_df, resources)
-
-    change_type_df = emission_calculator.get_change_type_table(emissions_df)
-    change_type_artifact = create_change_type_artifact(change_type_df, resources)
-
-    return [stock_artifact, summary_artifact, area_info_artifact, change_type_artifact]
-
-
-def create_chart_artifacts(
-    emissions_df: gpd.GeoDataFrame,
-    emissions_calculator: EmissionCalculator,
-    resources: ComputationResources,
-) -> List[Artifact]:
-    """
-    Contains the methods to create the chart artifacts.
-
-    :param emissions_df: Geodataframe with LULC change polygons and emissions [t] for each change type
-    :param emissions_calculator: Class containing the emission estimation methods
-    :param resources: Ephemeral computation resources
-    :return: List of the chart artifacts
-    """
-    area_chart_data = emissions_calculator.area_plot(emissions_df)
-    area_plot_artifact = create_area_plot_artifact(area_chart_data, resources)
-
-    emission_chart_data = emissions_calculator.emission_plot(emissions_df)
-    emission_plot_artifact = create_emission_plot_artifact(emission_chart_data, resources)
-
-    return [area_plot_artifact, emission_plot_artifact]